refactor(clubs): lift toQueryParamReader into a shared search-params module
Next hands a server component its query as a plain record while the param
parsers want a reader. That adapter is not specific to the club list, so it
moves to @/lib/search-params along with the QueryParamReader type. The club
list keeps its own parsing; only the shared adapter moves.
